scraper/requirements.txt
pandas==1.5.2
PyYAML==6.0
selenium==4.7.2
selenium_wire==5.1.0
webdriver_manager==3.8.5
undetected_chromedriver
openpyxl==3.0.10
pyautogui
requests